Further updates to migrate to new contract.

+++ Api/ApiModels/ChangeLink.cs
@@ -19,8 +19,8 @@ namespace Api.ApiModels
         public EntityName Name { get; set; }
         public string EntityId { get; set; }
         public EntityState State { get; set; }
-        public int DraftVersion { get; set; }
-        public int SubmittedVersion { get; set; }
+        public decimal DraftVersion { get; set; }
+        public decimal SubmittedVersion { get; set; }
         public string? Link { get; set; }
 
         public static ChangeLink Create(EntityBasics entityBasics, LinkGenerator linkGenerator, string customerId, string legalEntityId)
