+++ StateManagment/Models/EntityBasics.cs
@@ -5,7 +5,7 @@
         public EntityName Name { get; set; }
         public string EntityId { get; set; }
         public EntityState State { get; set; }
-        public int DraftVersion { get; set; }
-        public int SubmittedVersion { get; set; }
+        public decimal DraftVersion { get; set; }
+        public decimal SubmittedVersion { get; set; }
     }
 }
